Fix timing of postponed prompts
The time of the next prompt had incorrectly been calculated based on
the time of the first launch, which cannot be done reliably. Instead,
the time of the next prompt must be calculated based on the current
time. It could be set to half the previous interval, for example, or
simply to the original interval again, which is done here.